requirements.txt
flask
flask-cors
face_recognition
numpy
rembg
pillow
gunicorn